requirements.txt
faiss_cpu==1.11.0.post1
PyMuPDF==1.24.0
gradio==5.49.1
langchain==1.0.2
langchain_community==0.4
langchain_core==1.0.1
langchain_huggingface==1.0.0
langchain_openai==1.0.1
langchain_text_splitters==1.0.0
python-dotenv==1.2.1
python-docx==1.1.0
sentence_transformers==4.1.0
tqdm==4.67.1
